ai-service/requirements.txt
flask==3.0.0
tensorflow==2.15.0
pillow==10.2.0
numpy==1.26.3
opencv-python==4.9.0.80
pytesseract==0.3.10
scikit-learn==1.4.0
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0
redis==5.0.1
celery==5.3.6
pydantic==2.5.3
